C++ game server M5b-2 stage 1 part 3: the effect classes; passives and post-spawn skills go live
Three lanes, each with an adversarial reviewer and a fixer, then a regate
step. All seven gates pass with passive skills and npc post-spawn skills
live. The unit suite passes 2,914 of 2,915; the exception, tools.gen, was an
MSVC DLL-init crash under parallel load and passed alone. Lint is clean
over 3,659 files, and ownership is clean. The .cpp count of AION_UNPORTED
bodies falls from 1,852 to 1,747.

The effect classes (F-02, F-03, D13): the 38 classes of m5b2-plan.md
section 2.4 plus StumbleEffect, StaggerEffect, MPHealEffect and
ProcVPHealInstantEffect. That includes BufEffect, DamageEffect,
Heal/HealOverTime, the three masteries, Statup/Statdown, Root, Stun,
Shield, Hide, Provoker, Reflector, Sanctuary, Polymorph and Transform. Each
observer is registered through Effect::addObserver and each periodic task
through setPeriodicTask, as cycles.toml asks. tests/effects_al and the new
tests/effects_mz drive real Effects built from skill_templates.xml through
calculate, apply, start and end. The arithmetic classes are pinned by
golden values derived by hand from the Java.

The attack effect-engine half (F-06, D12): AttackUtil::calculateSkillResult,
calculateEffectResult, calculateMagicalOverTimeSkillResult and
calculatePhysicalStatus(effect), plus
StatFunctions::calculateMagicalSkillDamage. There are 19 golden-vector
cases, and 105 mutants were run against them.

The regate (G-05, D11): the O-09 and D7 holdbacks are removed. Every enter
world applies the character's passives, and startup casts the 199
post-spawn entries. Measured against the oracle: a fresh Warrior's
main-hand attack goes from 19/19 to 19/29 (weapon mastery 16% plus
statboost +7), and max HP/MP do not move. The M5b gate's kill falls from
about 12.6 s to 7.0 s. Two premises broke and are fixed:
- CheckOutput no longer requires zero live StatFunctionProxy. The npc
  statup buffs legitimately hold 1,171 of them at shutdown, which is the
  same precedent as KnownObject in M5b-1.
- The creation oracle now models the passive stat functions.
The OracleRunTest properties workaround is gone too, because the CMake
helper it avoided was fixed in part 2.

Known and carried forward:
- DuelService::fixTeamVisibility is still unported, so a Scout's Stealth
  (3222) leaves its hide and slow on the player. The fixup commit that
  follows ports it.
- The Drakenspire Depths post-spawn provokers reach the unported Bind and
  Silence effects inside instance 301390000.
- The RelWithDebInfo float run of risk 11 is still open.
- Five header requests are filed (m5b2-p3-1..5).

+++ cpp/game-server/src/aion/gameserver/skillengine/effect/AbstractOverTimeEffect.h
@@ -17,8 +17,8 @@ public:
 
 	int32_t getValue() const override { return value; }
 
-	/** on retail these effects last one sec more than their template value of duration2 */
-	int32_t getDuration2() const override { return duration2 + 1000; }
+	/** on retail these effects last one sec more than their template value of duration2 (Java int arithmetic: wraps above 2147482647) */
+	int32_t getDuration2() const override { return static_cast<int32_t>(static_cast<uint32_t>(duration2) + 1000u); }
 
 	void startEffect(model::Effect& effect) const override;
